# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "salus",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/goals/{goal_id}": {
      "delete": {
        "summary": "Api Delete Goal",
        "operationId": "api_delete_goal_api_v1_goals__goal_id__delete",
        "parameters": [
          {
            "name": "goal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Goal Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/insights/generate": {
      "post": {
        "summary": "Api Generate Insight",
        "operationId": "api_generate_insight_api_v1_insights_generate_post",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Date"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InsightResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/circadian/profile": {
      "post": {
        "summary": "Api Circadian Profile",
        "operationId": "api_circadian_profile_api_v1_circadian_profile_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CircadianProfileCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/_CircadianProfileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notifications/{notification_id}/read": {
      "post": {
        "summary": "Api Mark Notification Read",
        "operationId": "api_mark_notification_read_api_v1_notifications__notification_id__read_post",
        "parameters": [
          {
            "name": "notification_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Notification Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/onboarding/token": {
      "post": {
        "summary": "Api Onboarding Token",
        "operationId": "api_onboarding_token_api_v1_onboarding_token_post",
        "parameters": [
          {
            "name": "label",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "Webhook Token",
              "title": "Label"
            }
          },
          {
            "name": "scopes",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "ingest:write",
              "title": "Scopes"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/_OnboardingTokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/onboarding/entry": {
      "post": {
        "summary": "Api Onboarding Entry",
        "operationId": "api_onboarding_entry_api_v1_onboarding_entry_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/_OnboardingEntryBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/_OnboardingEntryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/onboarding/goal": {
      "post": {
        "summary": "Api Onboarding Goal",
        "operationId": "api_onboarding_goal_api_v1_onboarding_goal_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/_OnboardingGoalBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoalCreateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/goals/{item_id}": {
      "get": {
        "tags": [
          "goals"
        ],
        "summary": "Get One",
        "operationId": "get_one_api_v1_goals__item_id__get",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "goals"
        ],
        "summary": "Patch One",
        "operationId": "patch_one_api_v1_goals__item_id__patch",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "goals"
        ],
        "summary": "Delete One",
        "operationId": "delete_one_api_v1_goals__item_id__delete",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CircadianProfileCreate": {
        "properties": {
          "latitude": {
            "type": "number",
            "title": "Latitude"
          },
          "longitude": {
            "type": "number",
            "title": "Longitude"
          },
          "timezone_offset_hours": {
            "type": "number",
            "title": "Timezone Offset Hours",
            "default": 1.0
          },
          "configured_chronotype": {
            "type": "string",
            "title": "Configured Chronotype",
            "default": "intermediate"
          }
        },
        "type": "object",
        "required": [
          "latitude",
          "longitude"
        ],
        "title": "CircadianProfileCreate"
      },
...
      "GoalCreateResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "metric_code": {
            "type": "string",
            "title": "Metric Code"
          },
          "target_value": {
            "type": "number",
            "title": "Target Value"
          },
          "direction": {
            "type": "string",
            "title": "Direction"
          },
          "frequency": {
            "type": "string",
            "title": "Frequency"
          },
          "deadline": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline"
          }
        },
        "type": "object",
        "required": [
          "id",
          "metric_code",
          "target_value",
          "direction",
          "frequency"
        ],
        "title": "GoalCreateResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InsightResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "date": {
            "type": "string",
            "title": "Date"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "model_used": {
            "type": "string",
            "title": "Model Used"
          }
        },
        "type": "object",
        "required": [
          "id",
          "date",
          "content",
          "model_used"
        ],
        "title": "InsightResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "_CircadianProfileResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "latitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "timezone_offset_hours": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timezone Offset Hours"
          },
          "configured_chronotype": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Configured Chronotype"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id"
        ],
        "title": "_CircadianProfileResponse"
      },
...
      "_OnboardingEntryBody": {
        "properties": {
          "metric_code": {
            "type": "string",
            "title": "Metric Code"
          },
          "value": {
            "type": "string",
            "title": "Value"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "metric_code",
          "value"
        ],
        "title": "_OnboardingEntryBody"
      },
      "_OnboardingEntryResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "metric_code": {
            "type": "string",
            "title": "Metric Code"
          },
          "value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          },
          "timestamp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "id",
          "metric_code"
        ],
        "title": "_OnboardingEntryResponse"
      },
      "_OnboardingGoalBody": {
        "properties": {
          "metric_code": {
            "type": "string",
            "title": "Metric Code"
          },
          "target_value": {
            "type": "number",
            "title": "Target Value"
          },
          "direction": {
            "type": "string",
            "title": "Direction",
            "default": "increase"
          }
        },
        "type": "object",
        "required": [
          "metric_code",
          "target_value"
        ],
        "title": "_OnboardingGoalBody"
      },
      "_OnboardingTokenResponse": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token"
          },
          "webhook_url": {
            "type": "string",
            "title": "Webhook Url"
          }
        },
        "type": "object",
        "required": [
          "token",
          "webhook_url"
        ],
        "title": "_OnboardingTokenResponse"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}